requirements.txt
protobuf==5.28.0
pyyaml
requests